requirements.txt
torch>=2.1.0
torchvision>=0.16.0
pillow>=10.0.0
pandas>=2.2.0
scikit-learn>=1.3.0
tqdm>=4.66.0
requests>=2.31.0
fastapi>=0.110.0
uvicorn>=0.27.0
python-multipart>=0.0.9
pyyaml>=6.0.1